core/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
docker==7.1.0
psutil==6.1.1
pyyaml==6.0.1
python-multipart==0.0.9
websockets==12.0
aiofiles==23.2.1
python-dotenv==1.0.1
